+++ src/js/Form.js
@@ -924,7 +924,7 @@ Form.getRequiredTransformerVersion = function() {
     console.deprecate( 'Form.getRequiredTransformerVersion()', 'Form.requiredTransformerVersion' );
     return Form.requiredTransformerVersion;
 };
-Form.requiredTransformerVersion = '1.18.0';
+Form.requiredTransformerVersion = '1.19.0';
 
 module.exports = Form;