+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+----------------------
+##### Addeds
+- Support for appearance="numbers" on text inputs.
+
 [4.38.2] - 2017-08-17
 ----------------------
 ##### Fixed